Digit Bounding Box Detection.
Added the functionality to detected individual digits in an image.
Bounding boxes are generated using a Breath First Search algorithm.
Images are loaded using STB image.

The new functionality can be tested using the new VisTool.
This tool is written in JavaScript and uses the HTML5 canvas to draw stuff.

The maths library was expanded with Vector2, Rectangle and Functions classes.

Main.cpp got updated to have separate functions for testing the Neural Network and the Bounding Box features.