Implemented softmax & changed activation functions to jacobians
To implement softmax, the hadamard product had to be removed from the
backpropagation algorithm, and it needed to be replaced by
multiplication with a jacobian matrix. The other activation functions
had to be changed to output jacobians as well, so they work with the new
algorithm.
It was also necessary to add the Matrix.TotalSum() and
Matrix.LargestElem() functions to the Matrix class.

Minor change: placed network testing code inside the epoch loop in
Main.cpp, to get an evaluation every epoch, so you can see how the
network evolves.