Improved performance with a factor of 4.
Implemented an Array2D class which emulates a 2d array.
Internally it uses an std::vector to allocate all its elements in one allocation.
This saves a lot of allocations and gives a performance improvement with a factor of 4.
Indexing with rows and columns is emulated and thus a tiny bit slower than direct indexing.

The Matrix class was updated to use the new Array2D class instead of std::vector.
Multiple functions of the Matrix class were rewritten to use the direct indexing capabilities of the Array2D.

The member variables of the Matrix class are now private.
New getters to access these members were implemented and code that accessed the members got updated to use the new getters.