Datasets
Datasets allow us to make a quick selection of the images we want to use when training or testing a network.
They are a replacement for the old MNIST class but are almost entirely binary compatible.
The new implementation also gives a massive RAM memory improvement by only converting uint8 to double when an image is requested.

Solved a bug in the Matrix class that assumed the amount of classifications possible is always 10.

Some minor changes were made in the directory structure and some files were altered.