Implemented the Symbol Extractor class.
This class is responsible for extracting the individual symbols out of an image. It does this using some image processing and connected-component labelling to find the bounding boxes of the symbols. Finally, it converts the images into a format that is like the images the network was trained with.

Various tiny fixes where performed across the entire codebase. Both to make the development of the Symbol Extractor class easier and to improve the overall code quality.