Implemented a binary Dataset.
The binary dataset can be used to train networks to convert binary to decimal numbers.
This change was possible by making the Dataset class an abstract base class for providing training and testing data to a Network.
By inheriting from this class, all sorts of data sets can be made with ease.
The examples got updated to use the DatasetSymbols class for training with images of handwritten symbols.